fix: make TV timeline arithmetic explicit for CI
Xcode 26.3 rejected the existing mixed CGFloat/Double puck-offset expression as ambiguous, while the local Xcode build accepted it. Convert the progress fraction explicitly to CGFloat without changing the calculation, dimensions, focus or playback path. The same expression was present on the previous main revision.

Validation: diagnosed from the failed tvOS job in https://github.com/blurbery/vivid/actions/runs/34208670777. Remote revalidation is pending. This is a compiler-compatibility correction; the user-tested UI calculation and values are preserved.

AI-assisted with Codex for CI diagnosis and the explicit type conversion. I directed the overall work and confirmed the preceding device build. Model label unavailable.

+++ iosApp/iosApp/Screens/Player/tvOS/TVPlayerScrubber.swift
@@ -243,7 +243,7 @@ struct TVPlayerScrubber: View {
                 Ellipse().stroke(Color.white.opacity(isTimelineScrubbing ? 0.55 : 0), lineWidth: 1)
             )
             .shadow(color: .black.opacity(0.48), radius: 8, y: 3)
-            .offset(x: barWidth * progressFraction - (size + 10) / 2)
+            .offset(x: barWidth * CGFloat(progressFraction) - (size + 10) / 2)
     }
 
     private var trackHeight: CGFloat {
